#213 - lib: rethink native symbols
