+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mu"
-version = "0.1.52"
+version = "0.1.53"
 edition = "2021"
 authors = ["me <putnamjm.design@gmail.com>"]
