+++ metrics/regression/current.summary
@@ -1,40 +1,40 @@
 Perf Metrics Summary: lib
 --------------------
-lib/compile        bytes:     1520     usecs:     134.88
-lib/core           bytes:     5808     usecs:     458.82
-lib/gcd            bytes:      624     usecs:     320.58
-lib/list           bytes:     5296     usecs:     228.88
-lib/namespace      bytes:      240     usecs:       4.28
-lib/number         bytes:     3600     usecs:     162.66
-lib/reader         bytes:     5280     usecs:     288.96
-lib/special-form   bytes:     2400     usecs:     184.82
-lib/stream         bytes:      480     usecs:      34.86
-lib/struct         bytes:      576     usecs:      45.28
-lib/symbol         bytes:     1200     usecs:      91.94
-lib/vector         bytes:     4456     usecs:     260.06
+lib/compile        bytes:     1520     usecs:     103.50
+lib/core           bytes:     5808     usecs:     473.60
+lib/gcd            bytes:      624     usecs:     399.12
+lib/list           bytes:     5296     usecs:     332.52
+lib/namespace      bytes:      240     usecs:      12.54
+lib/number         bytes:     3600     usecs:     179.64
+lib/reader         bytes:     5280     usecs:     252.84
+lib/special-form   bytes:     2400     usecs:     158.22
+lib/stream         bytes:      480     usecs:      19.48
+lib/struct         bytes:      576     usecs:      34.14
+lib/symbol         bytes:     1200     usecs:      49.98
+lib/vector         bytes:     4456     usecs:     217.64
 
 Perf Metrics Summary: frequent
 --------------------
-frequent/core      bytes:     9624     usecs:    1599.48
-frequent/fixnum    bytes:     1680     usecs:     121.52
-frequent/float     bytes:     1200     usecs:      81.04
-frequent/list      bytes:     2160     usecs:     124.96
-frequent/vector    bytes:      240     usecs:       8.92
+frequent/core      bytes:     9624     usecs:     932.06
+frequent/fixnum    bytes:     1680     usecs:      81.48
+frequent/float     bytes:     1200     usecs:      78.40
+frequent/list      bytes:     2160     usecs:     126.98
+frequent/vector    bytes:      240     usecs:      13.06
 
 Perf Metrics Summary: prelude
 --------------------
-prelude/closure    bytes:    20904     usecs:   14071.94
-prelude/compile    bytes:     5512     usecs:    1559.44
-prelude/prelude    bytes:     4592     usecs:     332.86
-prelude/exception  bytes:     6896     usecs:    5107.96
-prelude/fixnum     bytes:     2000     usecs:     231.98
-prelude/format     bytes:     6152     usecs:    6891.54
-prelude/lambda     bytes:    97784     usecs:   65332.26
-prelude/list       bytes:    20864     usecs:    9017.64
-prelude/macro      bytes:    58416     usecs:   43679.58
-prelude/reader     bytes:    15032     usecs:  115693.64
-prelude/stream     bytes:      960     usecs:      83.44
-prelude/string     bytes:     2160     usecs:     630.06
-prelude/type       bytes:     8768     usecs:    6261.44
-prelude/vector     bytes:     9472     usecs:    6649.28
+prelude/closure    bytes:    20904     usecs:   15466.98
+prelude/compile    bytes:     5512     usecs:    1691.48
+prelude/prelude    bytes:     4592     usecs:     356.36
+prelude/exception  bytes:     6896     usecs:    5621.72
+prelude/fixnum     bytes:     2000     usecs:     247.06
+prelude/format     bytes:     6152     usecs:    7612.28
+prelude/lambda     bytes:    97784     usecs:   71217.84
+prelude/list       bytes:    20864     usecs:    9819.34
+prelude/macro      bytes:    58416     usecs:   48579.60
+prelude/reader     bytes:    15032     usecs:  127146.58
+prelude/stream     bytes:      960     usecs:      92.60
+prelude/string     bytes:     2160     usecs:     678.30
+prelude/type       bytes:     8768     usecs:    6753.48
+prelude/vector     bytes:     9472     usecs:    7230.64
