+++ metrics/regression/perf.l
@@ -42,7 +42,7 @@
                  (lib:car loop)
                  (lib:cdr loop))))
          (lib:cons () 1)))
-     (lib:hp-stat)
+     (lib:heap-stat)
      ((:lambda ()
          (lib:apply fn args)
-         (lib:hp-stat))))))
+         (lib:heap-stat))))))
