+++ src/libenv/core/mod.rs
@@ -11,7 +11,6 @@ pub mod dynamic;
 pub mod env;
 pub mod exception;
 pub mod frame;
-pub mod functions;
 pub mod future;
 pub mod gc;
 pub mod heap;
@@ -22,5 +21,6 @@ pub mod namespace;
 pub mod qquote;
 pub mod reader;
 pub mod readtable;
+pub mod symbols;
 pub mod types;
 pub mod utime;
