+++ src/prelude/inspect.l
@@ -8,7 +8,7 @@
    (:lambda (obj)
       ((:lambda (view)
           ((:lambda (tags)
-             (lib:struct :inspect `(function ,(lib:hp-size obj) ,tags)))
+             (lib:struct :inspect `(function ,(lib:heap-size obj) ,tags)))
            `(,(lib:cons :arity (lib:sv-ref view 0))
              ,(lib:cons :body (lib:sv-ref view 1)))))
        (lib:view obj))))
@@ -17,7 +17,7 @@
    (:lambda (obj)
      ((:lambda (view)
          ((:lambda (tags)
-            (lib:struct :inspect `(:<type> ,(lib:hp-size obj) ,tags)))
+            (lib:struct :inspect `(:<type> ,(lib:heap-size obj) ,tags)))
           `(,(lib:cons :type (lib:sv-ref (lib:st-vec obj) 0))
             ,(lib:cons :values (lib:sv-ref (lib:st-vec obj) 1)))))
       (lib:view obj))))
@@ -26,7 +26,7 @@
    (:lambda (obj)
      ((:lambda (view)
          ((:lambda (tags)  
-            (lib:struct :inspect `(struct ,(lib:hp-size obj) ,tags)))
+            (lib:struct :inspect `(struct ,(lib:heap-size obj) ,tags)))
           `(,(lib:cons :type (lib:st-type obj))
             ,(lib:cons :values (lib:st-vec obj)))))
       (lib:view obj))))
@@ -35,7 +35,7 @@
    (:lambda (obj)
      ((:lambda (view)
          ((:lambda (tags)  
-            (lib:struct :inspect `(symbol ,(lib:hp-size obj) ,tags)))
+            (lib:struct :inspect `(symbol ,(lib:heap-size obj) ,tags)))
           `(,(lib:cons :ns (lib:sv-ref view 0))
             ,(lib:cons :name (lib:sv-ref view 1))
             ,(lib:cons :value (lib:sv-ref view 2)))))
@@ -43,19 +43,19 @@
 
 (lib:intern :prelude "%inspect-const"
    (:lambda (obj)
-     (lib:struct :inspect `(,(prelude:type-of obj) ,(lib:hp-size obj) ()))))
+     (lib:struct :inspect `(,(prelude:type-of obj) ,(lib:heap-size obj) ()))))
 
 (lib:intern :prelude "%inspect-vector"
    (:lambda (obj)
      ((:lambda (tags)          
-         (lib:struct :inspect `(vector ,(lib:hp-size obj) ,tags)))
+         (lib:struct :inspect `(vector ,(lib:heap-size obj) ,tags)))
       `(,(lib:cons :type (lib:sv-type obj))
         ,(lib:cons :length (lib:sv-len obj))))))
 
 (lib:intern :prelude "%inspect-cons"
    (:lambda (obj)
      ((:lambda (tags)          
-         (lib:struct :inspect `(cons ,(lib:hp-size obj) ,tags)))
+         (lib:struct :inspect `(cons ,(lib:heap-size obj) ,tags)))
       `(,(lib:cons :car (lib:car obj))
         ,(lib:cons :cdr (lib:cdr obj))))))
