+++ src/prelude/time.l
@@ -43,10 +43,10 @@
                       (lib:car loop)
                       (lib:cdr loop))))
              '(() . 1))))
-        (lib:hp-stat)
+        (lib:heap-stat)
         ((:lambda (start)
            ((:lambda (value)
               `(,(lib:fx-sub (lib:proc-tm) start) ,@value))
             (prelude:apply fn args)))
         (lib:proc-tm))
-       (lib:hp-stat))))
+       (lib:heap-stat))))
