+++ tests/lib/core
@@ -14,7 +14,7 @@
 (lib:type-of 1)	:fixnum
 (lib:type-of :typeof)	:keyword
 (lib:type-of lib:std-in)	:stream
-(lib:type-of lib:hp-info)	:func
+(lib:type-of lib:heap-info)	:func
 (lib:type-of "type-of")	:vector
 (lib:view ())	#(:t :nil nil :nil)
 (lib:view :keyword)	#(:t :keyword keyword :keyword)
@@ -28,22 +28,22 @@
 (lib:view #s(:foo 1 2))	#(:t :foo #(:t 1 2))
 (lib:view 'a)	#(:t :nil a :UNBOUND)
 (lib:view 'lib:eq)	#(:t :lib eq #<:function :lib [req:2, form:eq]>)
-(lib:view lib:eq)	#(:t 2 :eq)
+(lib:view lib:eq)	#(:t 2 #(:t :lib eq 0))
 (lib:view lib:std-in)	#(:t 0 :input :nil)
-(lib:hp-size ())	8
-(lib:hp-size :keyword)	8
-(lib:hp-size #\a)	8
-(lib:hp-size '(1 2))	8
-(lib:hp-size 1234)	8
-(lib:hp-size 1.0)	8
-(lib:hp-size #s(:foo 1 2))	52
-(lib:hp-size "abcdefg")	8
-(lib:hp-size "abcdefgh")	26
-(lib:hp-size #(:t 1 2))	34
-(lib:hp-size 'a)	27
-(lib:hp-size 'lib:eq)	45
-(lib:hp-size lib:eq)	18
-(lib:hp-size lib:std-in)	8
+(lib:heap-size ())	8
+(lib:heap-size :keyword)	8
+(lib:heap-size #\a)	8
+(lib:heap-size '(1 2))	8
+(lib:heap-size 1234)	8
+(lib:heap-size 1.0)	8
+(lib:heap-size #s(:foo 1 2))	52
+(lib:heap-size "abcdefg")	8
+(lib:heap-size "abcdefgh")	26
+(lib:heap-size #(:t 1 2))	34
+(lib:heap-size 'a)	27
+(lib:heap-size 'lib:eq)	45
+(lib:heap-size lib:eq)	18
+(lib:heap-size lib:std-in)	8
 (lib:unwind (:lambda (cond obj) (lib:write cond () lib:std-out)) (:lambda () (lib:fx-div 1 1)))	1
 (lib:repr :t (lib:repr :vector :t))	:t
 (lib:type-of lib:unbound)	:func
